Added Automatic Server Backups & Camera Automations
Added an automatic server backup system.  * https://github.com/carstenschroeder/hassio-addons/

Added an automation that collects image captures in the front of the house.  Searches for humans, trucks, cars, cats, dogs.

Added an automation that turns the cameras off if the temperature drops below 16 degrees.  To lengthen the life.  These devices report that they cannot operate below 14 degrees.

Added an automation that turns the cameras back on when rising above 20 degrees if off.